Restructured files and fixed minor bugs
