+++ src/ADT/q_dinerdash.c
