--- src/dinerdash/boolean.h
@@ -1,8 +1,0 @@
-/*File : boolean.h*/
-
-#ifndef boolean_H
-#define boolean_H
-#define true 1
-#define false 0
-#define boolean unsigned char
-#endif
